frame_decode

Module extrinsics

This module contains functions for decoding extrinsics.

See decode_extrinsic for a general function to decode modern or historic extrinsics.

See decode_extrinsic_current for a helper to decode modern extrinsics.

See decode_extrinsic_legacy for a helper (and example) on decoding legacy extrinsics.

Structs

Extrinsic

Information about the extrinsic.

ExtrinsicCallInfo

Extrinsic call data information.

ExtrinsicExtensionInfo

Extrinsic extension information.

ExtrinsicExtensions

Information about the extrinsic signed extensions.

ExtrinsicInfoArg

An argument with a name and type ID.

ExtrinsicSignature

Information about the extrinsic signature.

ExtrinsicSignatureInfo

Extrinsic signature information.

Enums

ExtrinsicDecodeError

An error returned trying to decode extrinsic bytes.

ExtrinsicInfoError

An error returned trying to access extrinsic type information.

ExtrinsicType

The type of the extrinsic.

Traits

ExtrinsicTypeInfo

This is implemented for all metadatas exposed from frame_metadata and is responsible for extracting the type IDs that we need in order to decode extrinsics.

Functions

decode_extrinsic

Decode an extrinsic, returning information about it.

decode_extrinsic_current

Decode an extrinsic in a modern runtime (ie one exposing V14+ metadata).

decode_extrinsic_legacy

Decode an extrinsic in a historic runtime (ie one prior to V14 metadata). This is basically just an alias for decode_extrinsic.

Type Aliases

ExtrinsicOwned

An owned variant of an Extrinsic (note: this may still contain references if the visitor used to decode the extrinsic contents holds onto any)