frame_decode

Module storage

This module contains functions for decoding storage keys and values.

See decode_storage_key and decode_storage_value to decode storage keys or values from modern or historic runtimes.

See decode_storage_key_current and decode_storage_value_current to decode modern storage keys and values.

See [decode_storage_key_legacy] and [decode_storage_value_legacy] to decode historic storage keys and values (with examples).

Structs

StorageInfo

Information about a storage entry.

StorageKey

Details about a storage key.

StorageKeyInfo

Information about a single key within a storage entry.

StorageKeyPart

The decoded representation of a storage key.

StorageKeyPartValue

Information about the value contained within a storage key part hash.

Enums

StorageHasher

Hasher used by storage maps

StorageInfoError

An error returned trying to access storage type information.

StorageKeyDecodeError

An error returned trying to decode storage bytes.

StorageValueDecodeError

An error returned trying to decode storage bytes.

Traits

StorageTypeInfo

This is implemented for all metadatas exposed from frame_metadata and is responsible for extracting the type IDs and related info needed to decode storage entries.

Functions

decode_storage_key

Decode a storage key, returning information about it.

decode_storage_key_current

Decode a storage key in a modern runtime, returning information about it.

decode_storage_value

Decode a storage value.

decode_storage_value_current

Decode a storage value in a modern (V14-metadata-or-later) runtime.