frame_decode::storage

Trait StorageTypeInfo

pub trait StorageTypeInfo {
    type TypeId;

    // Required method
    fn get_storage_info(
        &self,
        pallet_name: &str,
        storage_entry: &str,
    ) -> Result<StorageInfo<Self::TypeId>, StorageInfoError<'_>>;
}

This is implemented for all metadatas exposed from frame_metadata and is responsible for extracting the type IDs and related info needed to decode storage entries.

Required Associated Types

type TypeId

The type of type IDs that we are using to obtain type information.

Required Methods

fn get_storage_info( &self, pallet_name: &str, storage_entry: &str, ) -> Result<StorageInfo<Self::TypeId>, StorageInfoError<'_>>

Get the information needed to decode a specific storage entry key/value.